requirement.txt
astroid==3.1.0
bcrypt==4.1.2
blinker==1.7.0
cachelib==0.12.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
dill==0.3.8
elastic-transport==8.12.0
elasticsearch==8.12.1
Flask==3.0.2
Flask-JWT-Extended==4.6.0
flask-msearch==0.2.9.4
Flask-MySQLdb==2.0.0
Flask-Session==0.6.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
mysqlclient==2.2.4
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pycparser==2.21
PyJWT==2.8.0
pylint==3.1.0
PyMySQL==1.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.28
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
Whoosh==2.7.4
WTForms==3.1.2
